Build final

+++ Assets/Scripts/SlotsMenu.cs
@@ -9,20 +9,46 @@ public class SlotsMenu : MonoBehaviour
     public GameObject Button;
     public GameObject TelaCarregamento;
 
+    [SerializeField]
+    public Text _title;
+
     [HideInInspector]
     public SaveController saveManager;
 
+    public AudioSource SemSave;
+    public AudioSource SaveDeletado;
+    public AudioSource Carregando;
+
+    [SerializeField]
+    public AudioSource NovoJogo;
+
+    public AudioSource JogarOndeParou;
+
+    [SerializeField]
+    public AudioSource jogo;
     void Awake()
     {
         saveManager = GameObject.Find("Slots").GetComponent<SaveController>();
     }
 
-    [SerializeField]
-    private Text _title;
-
     void Start()
     {
-        _title.text = PlayerPrefs.GetString("save1");
+        if(MiniTutorial.final == false)
+        {
+            _title.text = PlayerPrefs.GetString("save1");
+        }
+        else
+        {
+            final(MiniTutorial.final);
+        }
+        if(_title.text != "Novo jogo")
+        {
+            jogo.clip = JogarOndeParou.clip;
+        }
+        else
+        {
+            jogo.clip = NovoJogo.clip;
+        }
     }
 
     void OnDisable()
@@ -32,19 +58,41 @@ public class SlotsMenu : MonoBehaviour
 
     public void salvar()
     {
+        jogo.clip = JogarOndeParou.clip;
         TelaCarregamento.SetActive(true);
         _title.text = "Save 1: dia " + DateTime.Now;
         StartCoroutine(espera());
     }
     IEnumerator espera()
     {
+        Carregando.Play();
         yield return new WaitForSeconds(1);
         SceneManager.LoadScene(2);
     }
     public void deletar()
     {
-        _title.text = "Novo jogo";
-        PlayerPrefs.DeleteKey(_title.text);
+        if(_title.text != "Novo jogo")
+        {
+            jogo.clip = NovoJogo.clip;
+            SaveDeletado.Play();
+            _title.text = "Novo jogo";
+            PlayerPrefs.DeleteKey(_title.text);
+        }
+        else
+        {
+            SemSave.Play();
+        }
+        
+    }
+
+    public void final(bool final)
+    {
+        if(final == true)
+        {
+            jogo.clip = NovoJogo.clip;
+            _title.text = "Novo jogo";
+            MiniTutorial.final = false;
+        }
     }
 
 }
